Convert contents config to TS
