+++ .gitignore
@@ -20,6 +20,5 @@ static/assets
 static/collections
 static/contents
 src/lib/data/catalog.js
-src/lib/data/contents.js
 src/gen-assets
 vite.config.js.timestamp*
