+++ src/lib/components/BottomNavigationBar.svelte
@@ -4,7 +4,7 @@
 <script>
     import { base } from '$app/paths';
     import config from '$assets/config';
-    import contents from '$lib/data/contents';
+    import contents from '$assets/contents';
     import { language, languageDefault, refs, s, theme } from '$lib/data/stores';
     import { gotoRoute } from '$lib/navigate';
