+++ src/lib/components/Sidebar.svelte
@@ -5,9 +5,9 @@ The sidebar/drawer.
 <script>
     import { resolve } from '$app/paths';
     import config from '$assets/config';
+    import contents from '$assets/contents';
     import nav_drawer_image from '$assets/images/nav_drawer.png';
     import nav_drawer_2x from '$assets/images/nav_drawer@2x.png';
-    import contents from '$lib/data/contents';
     import {
         direction,
         fontChoices,
