+++ src/routes/text/+page.svelte
@@ -1,6 +1,7 @@
 <script>
     import { page } from '$app/state';
     import config from '$assets/config';
+    import contents from '$assets/contents';
     import AudioBar from '$lib/components/AudioBar.svelte';
     import BookSelector from '$lib/components/BookSelector.svelte';
     import BookTabs from '$lib/components/BookTabs.svelte';
@@ -13,7 +14,6 @@
     import { showTextAppearance } from '$lib/components/TextAppearanceSelector.svelte';
     import TextSelectionToolbar from '$lib/components/TextSelectionToolbar.svelte';
     import { playStop, seekToVerse, updateAudioPlayer } from '$lib/data/audio';
-    import contents from '$lib/data/contents';
     import {
         analytics,
         audioActive,
